+++ src/main/kotlin/org/dropProject/dropProjectPlugin/assignmentComponents/ListTable.kt
@@ -2,7 +2,7 @@ package org.dropProject.dropProjectPlugin.assignmentComponents
 
 import com.intellij.ide.DataManager
 import com.intellij.openapi.actionSystem.DefaultActionGroup
-import com.intellij.openapi.actionSystem.ex.ActionUtil
+import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.ui.SimpleTextAttributes
 import com.intellij.ui.awt.RelativePoint
@@ -49,7 +49,7 @@ class ListTable(private val toolWindow: DropProjectToolWindow) : TableView<Assig
 
         } else {
             emptyText.appendLine("Login to see your Assignments", SimpleTextAttributes.LINK_PLAIN_ATTRIBUTES, { actionEvent ->
-                ActionUtil.invokeAction(MainLogin(toolWindow), (actionEvent.source as MouseEvent).component, "", null, null)
+                ActionManager.getInstance().tryToExecute(MainLogin(toolWindow), null, (actionEvent.source as MouseEvent).component, "", true)
             })
         }
         addMouseListener(object : MouseAdapter() {